mcp-test-servers/notekeeper/requirements.txt
fastmcp>=2.10.5
uvicorn[standard]
python-dotenv
python-jose
starlette
sqlalchemy>=2.0.41